new ktam system

+++ src/base.rs
@@ -3,7 +3,7 @@ use thiserror;
 pub type Point = (usize, usize);
 pub type NumTiles = u32;
 pub type NumEvents = u64;
-pub type Tile = u32;
+pub type Tile = usize;
 pub type Energy = f64;
 pub type Glue = u32;
 pub type CanvasLength = usize;
